i have no clue what i'm doing can someone please help me omg i'm crying rn :'(
how do you use this. what are we doing. how have I got here. what is my purpose in life. why must we all go through this when in the end we'll all die and eventually be forgotten. is god real? what's out there? are we the only ones? help.